Feature EE-558: Adding organization filter
Get the organization tree that BG Process organization has. This way I can have all parents organization along with the Level Number, so I can order it to search for BiConnection in ascending way, using a TreeMap with levelno as key. Added error and message when data destination is empty for current BiConnection. Added BG Process organization and BiConnection organization as parameters so that the Python script can use them as filters.